docker/scripts/requirements.txt
aiosignal==1.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
frozenlist==1.3.0
gql==3.0.0
graphql-core==3.2.0
idna==3.3
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==6.0.2
platformdirs==2.5.0
pycodestyle==2.8.0
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.8
wrapt==1.13.3
yarl==1.7.2
